tests/coreos-install: move from tests/ to a subdir
Currently both the coreos-install testing as well as other testing
scripts are all bundled together inside of the tests/ folder. This
moves the coreos-install testing into a seperate subdirectory to
more easily distinguish what pieces belong to each.
