+++ tests/coreos-install/install_test.go
@@ -20,10 +20,10 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/coreos/init/tests/register"
-	"github.com/coreos/init/tests/util"
+	"github.com/coreos/init/tests/coreos-install/register"
+	"github.com/coreos/init/tests/coreos-install/util"
 
-	_ "github.com/coreos/init/tests/registry"
+	_ "github.com/coreos/init/tests/coreos-install/registry"
 )
 
 var flagBinaryPath string
